+++ lib/ui/page/home/page/chat/message_field/view.dart
@@ -745,8 +745,13 @@ class MessageFieldView extends StatelessWidget {
                             : 0,
                         child: CloseButton(
                           key: const Key('RemovePickedFile'),
-                          onPressed: () =>
-                              c.attachments.removeWhere((a) => a.value == e),
+                          onPressed: () {
+                            if (e is LocalAttachment) {
+                              e.cancelUpload();
+                            }
+
+                            c.attachments.removeWhere((a) => a.value == e);
+                          },
                         ),
                       );
                     }),