+++ lib/ui/page/home/page/my_profile/welcome_field/controller.dart
@@ -271,11 +271,18 @@ class WelcomeFieldController extends GetxController {
         var attachment = LocalAttachment(file, status: SendingStatus.sending);
         attachments.add(MapEntry(GlobalKey(), attachment));
 
-        Attachment uploaded = await _chatService.uploadAttachment(attachment);
+        final Attachment? uploaded = await _chatService.uploadAttachment(
+          attachment,
+        );
 
         int index = attachments.indexWhere((e) => e.value.id == attachment.id);
         if (index != -1) {
-          attachments[index] = MapEntry(attachments[index].key, uploaded);
+          // If `Attachment` returned is `null`, then it was canceled.
+          if (uploaded == null) {
+            attachments.removeAt(index);
+          } else {
+            attachments[index] = MapEntry(attachments[index].key, uploaded);
+          }
         }
       } on UploadAttachmentException catch (e) {
         MessagePopup.error(e);